+++ CHANGELOG.md
@@ -138,7 +138,7 @@
 * Define `State` type alias
 * Define `Stateful` type alias
 * Define `Scenario` type alias
-* Define `Stateful` pattern synonym
+* Define `Statefully` pattern synonym
 * Define `Scenario` pattern synonym
 * Define `rewrap` expression
 * Define `/` operator